Feb 2023 Pre-Return

+++ NASA-CHAPEA-WeRadMon3/12_disDat.ino
@@ -1,6 +1,6 @@
 void showMeTheDigits(){
 
-  float dataNow;
+  int dataNow;
   byte unitMarker;
   
   if      (wxDataType == 1){
